+++ components/suggest/src/suggestion.rs
@@ -88,6 +88,8 @@ pub enum Suggestion {
         title: String,
         total_reviews: i64,
         url: String,
+        icon: Option<Vec<u8>>,
+        icon_mimetype: Option<String>,
         score: f64,
     },
 }
@@ -127,7 +129,8 @@ impl Suggestion {
             | Self::Wikipedia { url, .. }
             | Self::Amo { url, .. }
             | Self::Yelp { url, .. }
-            | Self::Mdn { url, .. } => Some(url),
+            | Self::Mdn { url, .. }
+            | Self::Fakespot { url, .. } => Some(url),
             _ => None,
         }
     }
@@ -158,6 +161,16 @@ impl Suggestion {
             | Self::Mdn { title, .. } => title,
             Self::Weather { .. } => "weather",
             Self::Fakespot { title, .. } => title,
+        }
+    }
+
+    pub fn icon_data(&self) -> Option<&[u8]> {
+        match self {
+            Self::Amp { icon, .. }
+            | Self::Wikipedia { icon, .. }
+            | Self::Yelp { icon, .. }
+            | Self::Fakespot { icon, .. } => icon.as_deref(),
+            _ => None,
         }
     }
 }